Added more dynamic imports for better code splitting. Reorganized the code for better readiblity.
