+++ frontend/src/helpers/miscellaneous.js
@@ -42,3 +42,16 @@ export const downloadURI = (uri, name) => {
   link.click()
   document.body.removeChild(link)
 }
+
+/*
+isNumber code courtesy of: Dan
+Stackoverflow Post: https://stackoverflow.com/questions/175739/how-can-i-check-if-a-string-is-a-valid-number
+Dan's Profile: https://stackoverflow.com/users/17121/dan
+*/
+export const isNumber = str => {
+  // we only process strings!
+  if (typeof str != 'string') return false
+  // use type coercion to parse the entirety of the string (parseFloat alone does not do this)
+  // and ensure strings of whitespace fail
+  return !isNaN(str) && !isNaN(parseFloat(str))
+}
