+++ frontend/src/store/mutations.js
@@ -1,15 +1,4 @@
-export const roundOff = (num, scale) => {
-  if (!('' + num).includes('e')) {
-    return +(Math.round(num + 'e+' + scale) + 'e-' + scale)
-  } else {
-    var arr = ('' + num).split('e')
-    var sig = ''
-    if (+arr[1] + scale > 0) {
-      sig = '+'
-    }
-    return +(Math.round(+arr[0] + 'e' + sig + (+arr[1] + scale)) + 'e-' + scale)
-  }
-}
+import { roundOff } from '@/helpers/miscellaneous'
 
 const setImageSources = (state, imageSources) => {
   state.imageSources = imageSources
